first_RAG/src/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
python-dotenv==1.1.0
aiofiles==24.1.0
langchain==0.3.24
langchain-community 
langchain-core
PyMuPDF==1.25.5
motor==3.7.0
#bson==0.5.10
pymongo
pydantic-setting

